Yang: remove uses clause handling, now part of sonic-yang-mgmt (PR #3814)
sonic-yang-mgmt now integrates uses clauses as if a uses clause didn't
exist. No need to have this complex handling in sonic-utilities.

Besides removal of the code, the only test case changes are the removal
of the 'grouping' clause as it is specific to 'uses' which is not relevant
and some minor reordering of expected responses.

Signed-off-by: Brad House (@bradh352)